express08 done

+++ userHandlers.js
@@ -62,6 +62,12 @@ const postUser = async (req, res) => {
 const updateUser = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
+
+    if (id !== req.payload.sub) {
+      res.sendStatus(403);
+      return;
+    }
+
     const { firstname, lastname, email, city, language, hashedPassword } =
       req.body;
 
@@ -84,6 +90,11 @@ const updateUser = async (req, res) => {
 const deleteUser = (req, res) => {
   const id = parseInt(req.params.id);
 
+  if (id !== req.payload.sub) {
+    res.sendStatus(403);
+    return;
+  }
+
   database
     .query("delete from users where id = ?", [id])
     .then(([result]) => {
@@ -99,10 +110,31 @@ const deleteUser = (req, res) => {
     });
 };
 
+const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
+  const { email } = req.body;
+
+  database
+    .query("select * from users where email = ?", [email])
+    .then(([users]) => {
+      if (users[0] != null) {
+        req.user = users[0];
+
+        next();
+      } else {
+        res.sendStatus(401);
+      }
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).send("Error retrieving data from database");
+    });
+};
+
 module.exports = {
   getUsers,
   getUserById,
   postUser,
   updateUser,
   deleteUser,
+  getUserByEmailWithPasswordAndPassToNext,
 };
